SESSION FEATURE added in MVC

+++ springmvc/src/main/java/com/jspiders/springmvc/controller/StudentController.java
@@ -2,53 +2,68 @@ package com.jspiders.springmvc.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 
 import com.jspiders.springmvc.pojo.StudentPOJO;
 import com.jspiders.springmvc.service.StudentService;
 
 @Controller
 public class StudentController {
-	
+
 	@Autowired
 	private StudentService service;
 
 	@GetMapping("/home")
-	public String home() {
+	public String home(@SessionAttribute(name = "login", required = false) StudentPOJO login, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		return "Home";
 	}
 
 	@PostMapping("/login")
-	public String login(@RequestParam String username,
-			@RequestParam String password, ModelMap map) {
+	public String login(HttpServletRequest request, @RequestParam String username, @RequestParam String password,
+			ModelMap map) {
 		StudentPOJO student = service.login(username, password);
 		if (student != null) {
+			HttpSession session = request.getSession();
+			session.setAttribute("login", student);
+			session.setMaxInactiveInterval(30);
 			return "Home";
 		}
 		map.addAttribute("msg", "Inavlid username or password..!!");
 		return "LoginForm";
 	}
-	
+
 	@GetMapping("/add")
-	public String add() {
+	public String add(@SessionAttribute(name = "login", required = false) StudentPOJO login, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		return "AddStudent";
 	}
-	
+
 	@PostMapping("/add")
-	public String addData(@RequestParam String name,
-			@RequestParam String email,
-			@RequestParam long contact,
-			@RequestParam String city,
-			@RequestParam String username,
-			@RequestParam String password,
-			ModelMap map) {
-		StudentPOJO student = service.add(name, email,
-				contact, city, username, password);
+	public String addData(@SessionAttribute(name = "login", required = false) StudentPOJO login,
+			@RequestParam String name, @RequestParam String email, @RequestParam long contact,
+			@RequestParam String city, @RequestParam String username, @RequestParam String password, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
+		StudentPOJO student = service.add(name, email, contact, city, username, password);
 		if (student != null) {
 			map.addAttribute("student", student);
 		} else {
@@ -56,15 +71,23 @@ public class StudentController {
 		}
 		return "AddStudent";
 	}
-	
+
 	@GetMapping("/search")
-	public String search() {
+	public String search(@SessionAttribute(name = "login", required = false) StudentPOJO login, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		return "SearchStudent";
 	}
-	
+
 	@PostMapping("/search")
-	public String searchData(@RequestParam int id,
-			ModelMap map) {
+	public String searchData(@SessionAttribute(name = "login", required = false) StudentPOJO login,
+			@RequestParam int id, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		StudentPOJO student = service.search(id);
 		if (student != null) {
 			map.addAttribute("student", student);
@@ -73,17 +96,25 @@ public class StudentController {
 		}
 		return "SearchStudent";
 	}
-	
+
 	@GetMapping("/update")
-	public String update(ModelMap map) {
+	public String update(@SessionAttribute(name = "login", required = false) StudentPOJO login, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		List<StudentPOJO> students = service.searchAll();
 		map.addAttribute("students", students);
 		return "UpdateStudent";
 	}
-	
+
 	@PostMapping("/update")
-	public String updateForm(@RequestParam int id,
-			ModelMap map) {
+	public String updateForm(@SessionAttribute(name = "login", required = false) StudentPOJO login,
+			@RequestParam int id, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		StudentPOJO student = service.search(id);
 		if (student != null) {
 			map.addAttribute("student", student);
@@ -94,36 +125,43 @@ public class StudentController {
 		}
 		return "UpdateStudent";
 	}
-	
+
 	@PostMapping("/updateData")
-	public String updateData(@RequestParam int id,
-			@RequestParam String name,
-			@RequestParam String email,
-			@RequestParam long contact,
-			@RequestParam String city,
-			@RequestParam String username,
-			@RequestParam String password,
-			ModelMap map) {
+	public String updateData(@SessionAttribute(name = "login", required = false) StudentPOJO login,
+			@RequestParam int id, @RequestParam String name, @RequestParam String email, @RequestParam long contact,
+			@RequestParam String city, @RequestParam String username, @RequestParam String password, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		StudentPOJO student = service.search(id);
 		if (student != null) {
-			service.update(id, name, email, contact,
-					city, username, password);
+			service.update(id, name, email, contact, city, username, password);
 			map.addAttribute("msg", "Student updated successfully..!!");
 			List<StudentPOJO> students = service.searchAll();
 			map.addAttribute("students", students);
 		}
 		return "UpdateStudent";
 	}
-	
+
 	@GetMapping("/remove")
-	public String remove(ModelMap map) {
+	public String remove(@SessionAttribute(name = "login", required = false) StudentPOJO login, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		List<StudentPOJO> students = service.searchAll();
 		map.addAttribute("students", students);
 		return "RemoveStudent";
 	}
-	
+
 	@PostMapping("/remove")
-	public String removeData(@RequestParam int id, ModelMap map) {
+	public String removeData(@SessionAttribute(name = "login", required = false) StudentPOJO login,
+			@RequestParam int id, ModelMap map) {
+		if (login == null) {
+			map.addAttribute("msg", "Login first to proceed..!!");
+			return "LoginForm";
+		}
 		StudentPOJO student = service.remove(id);
 		if (student != null) {
 			map.addAttribute("msg", "Student removed successfully");
@@ -136,7 +174,9 @@ public class StudentController {
 	}
 
 	@GetMapping("/logout")
-	public String logout() {
+	public String logout(HttpSession session, ModelMap map) {
+		session.invalidate();
+		map.addAttribute("msg", "Successfully logged out..!!");
 		return "LoginForm";
 	}
 }
